+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/openai
 
+## 3.0.28
+
+### Patch Changes
+
+- Updated dependencies [4024a3a]
+  - @ai-sdk/provider-utils@4.0.15
+
 ## 3.0.27
 
 ### Patch Changes