+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Patch Changes
 
+- Updated dependencies [4024a3a]
+  - ai@6.0.84
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@6.0.83
 
 ## 0.0.1